Move Can to a new module for arrow-smash
