+++ arrow-smash/data/gradle.properties
@@ -1,0 +1,6 @@
+# Maven publishing configuration
+POM_NAME=Arrow Smash Data
+POM_ARTIFACT_ID=arrow-smash-data
+POM_PACKAGING=jar
+# Build configuration
+kapt.incremental.apt=false
