+++ arrow-smash/data/src/test/kotlin/arrow/smash/CanTest.kt
@@ -1,4 +1,4 @@
-package arrow.mtl.test
+package arrow.smash
 
 import arrow.core.Invalid
 import arrow.core.Ior
@@ -22,35 +22,22 @@ import arrow.core.test.laws.MonoidLaws
 import arrow.core.test.laws.ShowLaws
 import arrow.core.test.laws.TraverseLaws
 import arrow.core.toT
-import arrow.mtl.Can
-import arrow.mtl.component1
-import arrow.mtl.component2
-import arrow.mtl.extensions.can.applicative.applicative
-import arrow.mtl.extensions.can.bicrosswalk.bicrosswalk
-import arrow.mtl.extensions.can.bifunctor.bifunctor
-import arrow.mtl.extensions.can.bitraverse.bitraverse
-import arrow.mtl.extensions.can.eq.eq
-import arrow.mtl.extensions.can.eqK.eqK
-import arrow.mtl.extensions.can.eqK2.eqK2
-import arrow.mtl.extensions.can.functor.functor
-import arrow.mtl.extensions.can.hash.hash
-import arrow.mtl.extensions.can.monad.monad
-import arrow.mtl.extensions.can.monoid.monoid
-import arrow.mtl.extensions.can.show.show
-import arrow.mtl.extensions.can.traverse.traverse
-import arrow.mtl.getLeftOrElse
-import arrow.mtl.getOrElse
-import arrow.mtl.leftOption
-import arrow.mtl.rightOption
-import arrow.mtl.test.generators.can
-import arrow.mtl.test.generators.genK
-import arrow.mtl.test.generators.genK2
-import arrow.mtl.toCan
-import arrow.mtl.toIor
-import arrow.mtl.toLeftCan
-import arrow.mtl.toRightCan
-import arrow.mtl.toValidated
-import arrow.mtl.toValidatedLeft
+import arrow.smash.extensions.can.applicative.applicative
+import arrow.smash.extensions.can.bicrosswalk.bicrosswalk
+import arrow.smash.extensions.can.bifunctor.bifunctor
+import arrow.smash.extensions.can.bitraverse.bitraverse
+import arrow.smash.extensions.can.eq.eq
+import arrow.smash.extensions.can.eqK.eqK
+import arrow.smash.extensions.can.eqK2.eqK2
+import arrow.smash.extensions.can.functor.functor
+import arrow.smash.extensions.can.hash.hash
+import arrow.smash.extensions.can.monad.monad
+import arrow.smash.extensions.can.monoid.monoid
+import arrow.smash.extensions.can.show.show
+import arrow.smash.extensions.can.traverse.traverse
+import arrow.smash.generators.can
+import arrow.smash.generators.genK
+import arrow.smash.generators.genK2
 import arrow.typeclasses.Eq
 import io.kotlintest.properties.Gen
 import io.kotlintest.properties.forAll
